+++ src/main/java/com/global/api/builders/FraudBuilder.java
@@ -27,9 +27,6 @@ public class FraudBuilder<TResult> extends SecureBuilder<TResult> {
         return (TResult) client.processFraud(this);
     }
 
-    /**
-     * @return void
-     */
     @Override
     public void setupValidations() {
         validations